+++ facebook-core/src/test/kotlin/com/facebook/AccessTokenTest.kt
@@ -19,8 +19,10 @@
  */
 package com.facebook
 
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.test.core.app.ApplicationProvider
 import com.facebook.internal.Utility
 import com.facebook.internal.Utility.arrayList
@@ -31,17 +33,33 @@ import org.json.JSONObject
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
+import org.mockito.Matchers
 import org.powermock.api.mockito.PowerMockito
 import org.powermock.api.support.membermodification.MemberModifier
 import org.powermock.core.classloader.annotations.PrepareForTest
 
-@PrepareForTest(Utility::class, FacebookSdk::class)
+@PrepareForTest(Utility::class, FacebookSdk::class, LocalBroadcastManager::class)
 class AccessTokenTest : FacebookPowerMockTestCase() {
+
+  private val mockTokenString = "A token of my esteem"
+  private val mockAppID = "1234"
+  private val mockUserID = "1000"
+
   @Before
   fun before() {
     MemberModifier.stub<Any>(
             PowerMockito.method(Utility::class.java, "awaitGetGraphMeRequestWithCache"))
-        .toReturn(JSONObject().put("id", "1000"))
+        .toReturn(JSONObject().put("id", mockUserID))
+
+    PowerMockito.mockStatic(FacebookSdk::class.java)
+    PowerMockito.`when`(FacebookSdk.isInitialized()).thenReturn(true)
+    PowerMockito.`when`(FacebookSdk.getApplicationId()).thenReturn(mockAppID)
+    PowerMockito.`when`(FacebookSdk.getApplicationContext())
+        .thenReturn(ApplicationProvider.getApplicationContext())
+    PowerMockito.mockStatic(LocalBroadcastManager::class.java)
+    val mockLocalBroadcastManager = PowerMockito.mock(LocalBroadcastManager::class.java)
+    PowerMockito.`when`(LocalBroadcastManager.getInstance(Matchers.isA(Context::class.java)))
+        .thenReturn(mockLocalBroadcastManager)
   }
 
   @Test
@@ -424,5 +442,159 @@ class AccessTokenTest : FacebookPowerMockTestCase() {
     Assert.assertEquals(applicationId, accessToken.applicationId)
     Assert.assertEquals(userId, accessToken.userId)
     Assert.assertEquals(dataAccessExpirationTime, accessToken.dataAccessExpirationTime)
+  }
+
+  @Test
+  fun `test access token creation with default graph domain`() {
+    val accessToken =
+        AccessToken(
+            mockTokenString, mockAppID, mockUserID, null, null, null, null, null, null, null)
+    Assert.assertEquals(accessToken.graphDomain, "facebook")
+  }
+
+  @Test
+  fun `test access token creation with Instagram graph domain`() {
+    val accessToken =
+        AccessToken(
+            mockTokenString,
+            mockAppID,
+            mockUserID,
+            null, // permissions
+            null, // declined permissions
+            null, // expired permissions
+            null, // token source
+            null, // expiration time
+            null, // last refresh
+            null, // data access expiration time
+            "instagram")
+    Assert.assertEquals(accessToken.graphDomain, "instagram")
+  }
+
+  @Test
+  fun `test is logged in with Instagram`() {
+    val instagramAccessToken =
+        AccessToken(
+            mockTokenString,
+            mockAppID,
+            mockUserID,
+            null, // permissions
+            null, // declined permissions
+            null, // expired permissions
+            null, // token source
+            null, // expiration time
+            null, // last refresh
+            null, // data access expiration time
+            "instagram")
+    val facebookAccessToken =
+        AccessToken(
+            mockTokenString,
+            mockAppID,
+            mockUserID,
+            null, // permissions
+            null, // declined permissions
+            null, // expired permissions
+            null, // token source
+            null, // expiration time
+            null, // last refresh
+            null, // data access expiration time
+            "facebok")
+
+    AccessToken.setCurrentAccessToken(instagramAccessToken)
+    Assert.assertTrue(AccessToken.isLoggedInWithInstagram())
+    AccessToken.setCurrentAccessToken(facebookAccessToken)
+    Assert.assertFalse(AccessToken.isLoggedInWithInstagram())
+  }
+
+  @Test
+  fun `test default token source for Facebook domain`() {
+    val token = getAccessTokenWithSpecifiedSource("facebook", null)
+    Assert.assertEquals(token.source, AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test application web source for Facebook domain`() {
+    val token =
+        getAccessTokenWithSpecifiedSource("facebook", AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+    Assert.assertEquals(token.source, AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test chrome custom tab source for Facebook domain`() {
+    val token = getAccessTokenWithSpecifiedSource("facebook", AccessTokenSource.CHROME_CUSTOM_TAB)
+    Assert.assertEquals(token.source, AccessTokenSource.CHROME_CUSTOM_TAB)
+  }
+
+  @Test
+  fun `test webview source for Facebook domain`() {
+    val token = getAccessTokenWithSpecifiedSource("facebook", AccessTokenSource.WEB_VIEW)
+    Assert.assertEquals(token.source, AccessTokenSource.WEB_VIEW)
+  }
+
+  @Test
+  fun `test default token source for Gaming domain`() {
+    val token = getAccessTokenWithSpecifiedSource("gaming", null)
+    Assert.assertEquals(token.source, AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test application web source for Gaming domain`() {
+    val token =
+        getAccessTokenWithSpecifiedSource("gaming", AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+    Assert.assertEquals(token.source, AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test chrome custom tab source for Gaming domain`() {
+    val token = getAccessTokenWithSpecifiedSource("gaming", AccessTokenSource.CHROME_CUSTOM_TAB)
+    Assert.assertEquals(token.source, AccessTokenSource.CHROME_CUSTOM_TAB)
+  }
+
+  @Test
+  fun `test webview source for Gaming domain`() {
+    val token = getAccessTokenWithSpecifiedSource("gaming", AccessTokenSource.WEB_VIEW)
+    Assert.assertEquals(token.source, AccessTokenSource.WEB_VIEW)
+  }
+
+  @Test
+  fun `test default token source for Instagram domain`() {
+    val token = getAccessTokenWithSpecifiedSource("instagram", null)
+    Assert.assertEquals(token.source, AccessTokenSource.INSTAGRAM_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test application web source for Instagram domain`() {
+    val token =
+        getAccessTokenWithSpecifiedSource("instagram", AccessTokenSource.FACEBOOK_APPLICATION_WEB)
+    Assert.assertEquals(token.source, AccessTokenSource.INSTAGRAM_APPLICATION_WEB)
+  }
+
+  @Test
+  fun `test chrome custom tab source for Instagram domain`() {
+    val token = getAccessTokenWithSpecifiedSource("instagram", AccessTokenSource.CHROME_CUSTOM_TAB)
+    Assert.assertEquals(token.source, AccessTokenSource.INSTAGRAM_CUSTOM_CHROME_TAB)
+  }
+
+  @Test
+  fun `test webview source for Instagram domain`() {
+    val token = getAccessTokenWithSpecifiedSource("instagram", AccessTokenSource.WEB_VIEW)
+    Assert.assertEquals(token.source, AccessTokenSource.INSTAGRAM_WEB_VIEW)
+  }
+
+  fun getAccessTokenWithSpecifiedSource(
+      graphDomain: String,
+      tokenSource: AccessTokenSource?
+  ): AccessToken {
+    return AccessToken(
+        mockTokenString,
+        mockAppID,
+        mockUserID,
+        null, // permissions
+        null, // declined permissions
+        null, // expired permissions
+        tokenSource,
+        null, // expiration time
+        null, // last refresh
+        null, // data access expiration time
+        graphDomain)
   }
 }